Updating to latest FlutterFlow output.

+++ lib/pages/meucarrinho/meucarrinho_widget.dart
@@ -533,7 +533,8 @@ class _MeucarrinhoWidgetState extends State<MeucarrinhoWidget>
                                                                     pedidosItem
                                                                         .precoBorda) *
                                                                 (pedidosItem
-                                                                    .quanty);
+                                                                    .quanty
+                                                                    .toDouble());
                                                       });
                                                     },
                                                     child: Icon(
